MINT-1728. Removed old ContractManager. Renamed ContractManager3DAO to ContractManagerDAO. Removed old AbstractContractDAO. Renamed AbstractContractDAO3 to AbstractContractDAO. Removed useless network checking. Removed unused DAOs: ERC20LibraryDAO, FakeCoinDAO, LHTDAO. Removed unused components.
