+++ config/webpack.config.dev.js
@@ -7,7 +7,7 @@ const path = require('path')
 const webpack = require('webpack')
 const babel = require('./babel.dev')
 const CompileTimePlugin = require('webpack-compile-time-plugin')
-// const CircularDependencyPlugin = require('circular-dependency-plugin')
+const CircularDependencyPlugin = require('circular-dependency-plugin')
 
 const config = require('./webpack.config.base.js')
 
@@ -35,22 +35,22 @@ module.exports = config.buildConfig(
     },
     babel,
     plugins: [
-      // new CircularDependencyPlugin({
-      //   // `onStart` is called before the cycle detection starts
-      //   onStart ({ compilation }) {
-      //     console.log('start detecting webpack modules cycles')
-      //   },
-      //   // `onDetected` is called for each module that is cyclical
-      //   onDetected ({ module: webpackModuleRecord, paths, compilation }) {
-      //     if (!/node_modules/.test(webpackModuleRecord.resource)) {
-      //       compilation.errors.push(new Error(paths.join(' -> ')))
-      //     }
-      //   },
-      //   // `onEnd` is called before the cycle detection ends
-      //   onEnd ({ compilation }) {
-      //     console.log('end detecting webpack modules cycles')
-      //   },
-      // }),
+      new CircularDependencyPlugin({
+        // `onStart` is called before the cycle detection starts
+        onStart ({ compilation }) {
+          console.log('start detecting webpack modules cycles')
+        },
+        // `onDetected` is called for each module that is cyclical
+        onDetected ({ module: webpackModuleRecord, paths, compilation }) {
+          if (!/node_modules/.test(webpackModuleRecord.resource)) {
+            compilation.errors.push(new Error(paths.join(' -> ')))
+          }
+        },
+        // `onEnd` is called before the cycle detection ends
+        onEnd ({ compilation }) {
+          console.log('end detecting webpack modules cycles')
+        },
+      }),
       new CompileTimePlugin(),
       new HtmlWebpackPlugin({
         inject: true,
