+++ package.json
@@ -30,6 +30,7 @@
     "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-stage-0": "^6.24.1",
     "chalk": "1.1.3",
+    "circular-dependency-plugin": "4.4.0",
     "copy-webpack-plugin": "^4.0.0",
     "coveralls": "^2.13.1",
     "cross-env": "^5.0.0",
