--- packages/core-dependencies/constants.js
@@ -1,9 +1,0 @@
-/**
- * Copyright 2017–2018, LaborX PTY
- * Licensed under the AGPL Version 3 license.
- *
- * @flow
- */
-/* eslint-disable import/prefer-default-export */
-
-export { FORM_CBE_TOKEN } from '../../src/components/constants'
