+++ packages/core/dao/AssetsManagerDAO.js
@@ -9,7 +9,7 @@ import { unionBy } from 'lodash'
 import OwnerCollection from '../models/wallet/OwnerCollection'
 import OwnerModel from '../models/wallet/OwnerModel'
 import BlacklistModel from '../models/tokens/BlacklistModel'
-import AbstractContractDAO from './AbstractContract3DAO'
+import AbstractContractDAO from './AbstractContractDAO'
 import ChronoBankPlatformDAO from './ChronoBankPlatformDAO'
 import ChronoBankAssetDAO from './ChronoBankAssetDAO'
 import PlatformsManagerDAO from './PlatformsManagerDAO'
