+++ packages/core/dao/MultisigWalletDAO.js
@@ -7,7 +7,7 @@ import { zipWith } from 'lodash'
 import BigNumber from 'bignumber.js'
 import resultCodes from 'chronobank-smart-contracts/common/errors'
 import AbstractMultisigContractDAO from './AbstractMultisigContractDAO'
-import AbstractContractDAO from '../dao/AbstractContract3DAO'
+import AbstractContractDAO from './AbstractContractDAO'
 import Amount from '../models/Amount'
 import TokenModel from '../models/tokens/TokenModel'
 import TxExecModel from '../models/TxExecModel'
