+++ packages/core/redux/session/thunks.js
@@ -11,10 +11,8 @@ import privateKeyProvider from '@chronobank/login/network/privateKeyProvider'
 import { push, replace } from '@chronobank/core-dependencies/router'
 import LocalStorage from '@chronobank/core-dependencies/utils/LocalStorage'
 import web3Provider from '@chronobank/login/network/Web3Provider'
-import * as Utils from '@chronobank/login/redux/network/utils'
 import setup from '@chronobank/login/network/EngineUtils'
 import metaMaskResolver from '@chronobank/login/network/metaMaskResolver'
-import contractsManagerDAO from '../../dao/ContractsManagerDAO'
 import * as SessionActions from './actions'
 import * as ProfileThunks from '../profile/thunks'
 import ProfileService from '../profile/service'
@@ -38,33 +36,6 @@ export const checkMetaMask = () => (dispatch) => {
   metaMaskResolver
     .on('resolve', metaMaskHandler)
     .start()
-}
-
-export const checkNetwork = () => async (dispatch) => {
-  dispatch(NetworkActions.loading())
-  const isDeployed = await contractsManagerDAO.isDeployed()
-  if (!isDeployed) {
-    dispatch(NetworkActions.addError('Network is unavailable.'))
-  }
-  //return isDeployed
-  return true
-}
-
-export const checkLocalSession = (account, providerURL) => async (dispatch) => {
-  const isTestRPC = Utils.checkTestRPC(providerURL)
-  // testRPC must be exists
-  if (!isTestRPC || !account) {
-    return false
-  }
-
-  // contacts and network must be valid
-  const isDeployed = await dispatch(checkNetwork())
-  if (!isDeployed) {
-    return false
-  }
-
-  // all tests passed
-  return true
 }
 
 export const getAccounts = () => (dispatch, getState) => {
@@ -182,8 +153,6 @@ export const logout = () => async (dispatch, getState) => {
     dispatch(push('/'))
     if (selectedNetworkId === NETWORK_MAIN_ID) {
       location.reload()
-    } else {
-      await dispatch(bootstrap(false))
     }
   } catch (e) {
     // eslint-disable-next-line
@@ -227,29 +196,8 @@ export const login = (account) => async (dispatch, getState) => {
   dispatch(replace(LocalStorage.getLastURL() || defaultURL))
 }
 
-export const bootstrap = (relogin = true, isMetaMaskRequired = true, isLocalAccountRequired = true) => async (dispatch, getState) => {
-  if (isMetaMaskRequired) {
-    dispatch(checkMetaMask())
-  }
-
-  if (!relogin) {
-    return true
-  }
-
-  if (isLocalAccountRequired) {
-    const localAccount = LocalStorage.getLocalAccount()
-    const isPassed = await dispatch(checkLocalSession(localAccount))
-    if (isPassed) {
-      await dispatch(restoreLocalSession(localAccount, getState().get('ethMultisigWallet')))
-      dispatch(createNetworkSession(localAccount, LOCAL_PROVIDER_ID, LOCAL_ID))
-      dispatch(login(localAccount))
-    } else {
-      // eslint-disable-next-line
-      console.warn('Can\'t restore local session')
-    }
-  }
-
-  return true
+export const bootstrap = () => async () => {
+  return true //FIXME remove method
 }
 
 export const getProfileSignature = (wallet) => async (dispatch) => {
