+++ packages/login/redux/network/thunks.js
@@ -7,7 +7,6 @@
 
 import * as PersistAccountActions from '@chronobank/core/redux/persistAccount/actions'
 import {
-  checkNetwork,
   createNetworkSession,
   getProviderSettings,
   login,
@@ -33,6 +32,7 @@ import {
   LOCAL_PRIVATE_KEYS,
   NETWORK_MAIN_ID,
 } from '../../network/settings'
+import { DUCK_ETH_MULTISIG_WALLET } from '@chronobank/core/redux/multisigWallet/constants'
 
 /*
  * Thunk dispatched by "" screen.
@@ -120,17 +120,12 @@ export const handleWalletLogin = (wallet, password) => async (dispatch, getState
 
   dispatch(NetworkActions.clearErrors())
 
-  const isPassed = await dispatch(checkNetwork())
-
-  if (isPassed) {
-    dispatch(createNetworkSession(
-      selectedAccount,
-      selectedProviderId,
-      selectedNetworkId,
-    ))
-    await dispatch(login(selectedAccount))
-  }
-
+  dispatch(createNetworkSession(
+    selectedAccount,
+    selectedProviderId,
+    selectedNetworkId,
+  ))
+  await dispatch(login(selectedAccount))
 }
 
 /*
@@ -151,7 +146,7 @@ export const handleLoginLocalAccountClick = (account = '') =>
   async (dispatch, getState) => {
     let state = getState()
     const { accounts } = state.get(DUCK_NETWORK)
-    const wallets = state.get('ethMultisigWallet') // FIXME: to use constant
+    const wallets = state.get(DUCK_ETH_MULTISIG_WALLET)
     const providerSetting = dispatch(getProviderSettings())
     const index = Math.max(accounts.indexOf(account), 0)
     const provider = privateKeyProvider.getPrivateKeyProvider(
@@ -171,17 +166,12 @@ export const handleLoginLocalAccountClick = (account = '') =>
 
     dispatch(NetworkActions.clearErrors())
 
-    // checkNetwork has no arguments. See the diff.
-    const isPassed = await dispatch(checkNetwork())
-
-    if (isPassed) {
-      dispatch(createNetworkSession(
-        selectedAccount,
-        selectedProviderId,
-        selectedNetworkId,
-      ))
-      dispatch(login(selectedAccount))
-    }
+    dispatch(createNetworkSession(
+      selectedAccount,
+      selectedProviderId,
+      selectedNetworkId,
+    ))
+    dispatch(login(selectedAccount))
   }
 
 /*
