+++ src/components/constants.js
@@ -9,7 +9,6 @@ export const FORM_ADD_PLATFORM_DIALOG = 'AddPlatformDialog'
 export const FORM_ADD_TOKEN_DIALOG = 'AddTokenDialog'
 export const FORM_ASSET_MANAGER = 'AssetManagerDialog'
 export const FORM_CBE_ADDRESS = 'CBEAddressDialog'
-export const FORM_CBE_TOKEN = 'CBETokenDialog'
 export const FORM_CREATE_EXCHANGE = 'createExchangeForm'
 export const FORM_CUSTOM_WALLET_ADD = 'CustomWalletForm'
 export const FORM_DEPOSIT_TOKENS = 'FormDepositTokens'
