+++ src/components/dashboard/SendTokens/SendTokensForm.jsx
@@ -9,7 +9,6 @@ import { BLOCKCHAIN_BITCOIN, BLOCKCHAIN_BITCOIN_CASH, BLOCKCHAIN_BITCOIN_GOLD, B
 import { TOKEN_ICONS } from 'assets'
 import Preloader from 'components/common/Preloader/Preloader'
 import TokenValue from 'components/common/TokenValue/TokenValue'
-import contractsManagerDAO from '@chronobank/core/dao/ContractsManagerDAO'
 import { TX_TRANSFER } from '@chronobank/core/dao/constants/ERC20DAO'
 import web3Converter from '@chronobank/core/utils/Web3Converter'
 import Amount from '@chronobank/core/models/Amount'
@@ -407,14 +406,6 @@ export default class SendTokensForm extends PureComponent {
     }
 
     return null
-  }
-
-  isFieldValid = (fieldName) => {
-    return fieldName in this.props.formErrors
-  }
-
-  checkIsContract (address): Promise {
-    return contractsManagerDAO.isContract(address)
   }
 
   convertSatoshiToBTC = (satoshiAmount) => {
