+++ src/specsInit.js
@@ -3,7 +3,6 @@
  * Licensed under the AGPL Version 3 license.
  */
 
-import resultCodes from 'chronobank-smart-contracts/common/errors'
 import Reverter from 'chronobank-smart-contracts/test/helpers/reverter'
 import customSerializer from '@chronobank/core-dependencies/utils/CustomSerializer'
 import Enzyme from 'enzyme'
@@ -15,7 +14,6 @@ import web3provider from '@chronobank/login/network/Web3Provider'
 import configureMockStore from 'redux-mock-store'
 import thunk from 'redux-thunk'
 import Web3 from 'web3'
-import AbstractContractDAO from '@chronobank/core/dao/AbstractContractDAO'
 import MarketSocket from '@chronobank/core/market/MarketSocket'
 import ls from '@chronobank/core-dependencies/utils/LocalStorage'
 
@@ -29,7 +27,8 @@ web3provider.reinit(web3, new web3.providers.HttpProvider('http://localhost:8545
 web3provider.resolve()
 export const accounts = web3.eth.accounts
 
-AbstractContractDAO.setup(accounts[0], [resultCodes.OK, true], resultCodes)
+// FIXME revive somehow
+// AbstractContractDAO.setup(accounts[0], [resultCodes.OK, true], resultCodes)
 
 const reverter = new Reverter(web3provider.getWeb3instance())
 
@@ -59,7 +58,6 @@ beforeEach(() => {
 
 afterEach(async (done) => {
   ls.destroySession()
-  await AbstractContractDAO.stopWholeWatching()
   done()
 })
